First set - Assignment 4.
This is a shell script to accept two filenames as arguments. First file will be having the username and the second file will be having the passwords for it. If the number of usernames doesn't match with number of passwords it will not create any users, if it matches it will create user with corresponding password.